feat(backend): aprsing country from demo + added uploader id
