+++ apps/frontend/src/pages/wr-database/components/wr-list/wr-list.component.html
@@ -64,7 +64,7 @@
             ></app-player-cell>
           } @else {
             <span class="df-name">
-              <app-flag [country]="null"></app-flag>
+              <app-flag [country]="record.playerCountry"></app-flag>
               {{ record.playerNick }}
             </span>
           }
