+++ apps/frontend/src/shared/interfaces/country.interface.ts
@@ -1,6 +1,1 @@
-export interface CountryInterface {
-  fullName: string;
-  shortName: string;
-  col: number;
-  row: number;
-}
+export { CountryInterface } from '@dfcomps/contracts';
