+++ apps/frontend/src/shared/modules/last-wrs/last-wrs.component.html
@@ -26,7 +26,7 @@
           ></app-player-cell>
         } @else {
           <div class="df-name">
-            <app-flag [country]="null"></app-flag>
+            <app-flag [country]="record.playerCountry"></app-flag>
             <div class="df-name-caption">{{ record.playerNick }}</div>
           </div>
         }
