+++ apps/frontend/src/shared/rest-api/business/url-params.config.ts
@@ -133,14 +133,14 @@ export class URL_PARAMS {
   public static get WORLD_RECORDS(): {
     UPLOAD: string;
     LIST: (page: number, filter: string, physics: string) => string;
-    LAST_FIVE: string;
+    LAST_FIVE: (physics: string) => string;
     SEARCH_PLAYERS: (nick: string) => string;
   } {
     return {
       UPLOAD: `${API_URL}/world-records/upload`,
       LIST: (page: number, filter: string, physics: string) =>
         `${API_URL}/world-records/list?page=${page}&filter=${encodeURIComponent(filter)}&physics=${physics}`,
-      LAST_FIVE: `${API_URL}/world-records/last-five`,
+      LAST_FIVE: (physics: string) => `${API_URL}/world-records/last-five?physics=${physics}`,
       SEARCH_PLAYERS: (nick: string) => `${API_URL}/world-records/search-players?nick=${encodeURIComponent(nick)}`,
     };
   }
