+++ libs/contracts/src/lib/global/country.interface.ts
@@ -1,0 +1,6 @@
+export interface CountryInterface {
+  fullName: string;
+  shortName: string;
+  col: number;
+  row: number;
+}
