+++ libs/contracts/src/lib/global/index.ts
@@ -2,3 +2,5 @@ export { CountInterface } from './count.interface';
 export { Physics } from './physics.enum';
 export { UploadedFileLinkInterface } from './uploaded-file-link.interface';
 export { Languages } from './languages.enum';
+export { CountryInterface } from './country.interface';
+export { COUNTRIES_CONFIG } from './countries.config';
